ENH: first draft of skeptical signal
This is a class for dealing with detectors which do not
have a trigger at the hardware level, but are being scanned
against anyway.  This adds a software trigger which requires that
at least one update from the IOC (which changes the timestamp) happen
before the detector reports it is done.

This is to account for IOCs which when encountering a bad
value chose to return 0 instead nan.

If the value coming out of a monitor subscription is 0.0, reject this
update and continue with life.  This short-circuits all of the base-class
logic which updates the python-side value + timestamp and runs the
ophyd layer callbacks.

FIX: Fix wrong signature