FIX: make Device.configuration_attrs include data
During an offline discussion, we found that if a component is listed in
the configuration_attrs list, we want both its metadata and its values.

For example, if component 'a' is in a device's read_configuration list,
calling device.describe_configuration() will return an aggregate of both
device.a.describe_configuration() and device.a.describe().