API: remove FileStoreBulkWrite
Now that the Registry is generating the datum UIDs the 'stash, emit on
read, and then insert on unstage' is no longer possible.  This means
we will never let a datum_id which is not in a Registry out into the
EventSources.

This change is driven by the need to support column based backends
from Assets.