API: do not call both read_configuration and read
For complex devices, the same child device may be used in both
read_attrs and read_configuration.  Putting the read values into the
configuration is generically not correct (ex, the mean_value of a
stats plugin for AD should be in the read() but not in the result of
read_configuration() ).

At the bottom Signals fall back to read() for their read_configuration
implementation.