API: make the name keyword to device a required keyword only arg
This ensures that the names that appear in the read dictionary are
always human-readable.

In practice if this is left out it is accidental and results in saved
data with keys that are a concatenation of the PV prefix with the
names of the children in the python class.  This is neither human
understandable nor a valid / existing PV.  Neither humans nor machines
can read this.

Keeping keyword only to preserve the ability to have sub-classes with
more than one required positional arguments.

Do not want to make it the first positional argument to avoid massive
API breakage and the convention of 'last positional argument' is
brittle.

A keyword-only argument also clearly disambiguates the name (for
humans) from the prefix (for hardware).