API/ENH: make read_attrs dive deep instead of only top-level
Include all generation in read_attrs / configuration_attrs

One thing that you can do with the current system that you can mix
direct child and grand (, grand ...)-child component names in
read_attrs and it does the right thing (because
getattr(grand_child).read() does what you expect). This 'generation
skipping' is convenient, but can lead to confusingly set up devices
where a child is not in it's parents read_attrs and it's read_attrs
are badly configured, but via this generation skipping one of it's
children is in the parents read_attrs and the top-level device mostly
does the right thing (however it is my suspicion that we are not
reading all of the configuration we should be).

We can support grand-children via the read_attrs.setter in
a (hopefully) non-controversial way be detecting them add then
re-cursing.  This has the effect of 'correctly' setting up the
read_attrs list of the children. The new system also ensures that
sub-devices are in their parents configuration_attrs if they are in
the read_attrs. This is a small API change, but arguably a bug-fix.

The only thing I think this would break is someone being much too
clever and using dev.read_attrs = ['grand.child'] dev.grand.read_attrs
= ['some', 'other', 'children'] and expecting dev.read() and
dev.grand.read() to have completely dis-joint keys. I don't really see
a use-case for that and the new behavior of forcing dev.grand.read()
being a strict sub-set of dev.read() makes a lot more sense.

So far so good, but a maybe bigger hitch is that we can not support
arbitrarily mixed-generation read_attrs (that is dev.read_attrs =
['child1', 'child2.a']). We have 3 choices we can support with the new
'truth lives on the components' scheme (but they all change behavior)

 1 only include direct children. This mean if the user does
   dev.read_attrs = ['grand.child'], then the accessed value would be
   `dev.read_attrs == ['grand']

 2 recuse down to the leaves and only include the end signals. In this
   case the length of read_attrs and the length of dev.read() are the
   same and [getattr(dev, n).name for n in dev.read_attrs] gets you
   back the read keys. However this means doing somethnig like
   dev.read_attrs = ['grand'] would result ind dev.read_attrs ==
   ['grand.child1', 'grand.child2']. This has the weird side effect
   that if no child devices have any configuration to offer, then
   dev.configuration_attrs is empty which is confusing.

 3 include all generations of names, basically combine 1 and 2.

This implements option 3, the initial implementation was 1.