FIX: handle np.nan as position
This comes up in the context of a pseudo-positioner at HXN that may be
in a real-motor configuration which does not map to a valid
pseudo-position.  We dealt with this by having the conversion function
return nans which in turn causes warnings to come out of the progress
bar when it can not convert the nan -> int.