FIX: ensure_blocking() as implemented was a no-op.
I believe V33CamMixin.ensure_blocking() was intended to loop over a
detector's plugins and call PluginBase.ensure_nonblocking() on each. As
written, this loops over the components of Cam, whereas the plugins are
compoents of a Cam's parent, the AreaDetector itself.

This would run fine, but it will never find the plugins it is looking
for.