ENH: Add register_instance.
At various times it has been proposed to "auto-register" ophyd objects
in order to maintain a global registery of devices. We have avoided
making OphydObj aware of all its own instances because this could have
weird implications and generally feels like an unusual design in Python.
But we could allow downstream code to sign up to be notified each time
an OphydObj is instantiated. Each subscriber will get one notification,
and they can keep a strong or weak reference if they like, as needed.

<hr />

This suggestion is fully due to @tacaswell and I may or may not have
implemented it exactly as he envisioned. I need to refer to it in
another discussion so I have sketched it here.