MNT: remove all of the re-try logic
After discussion and experimentation we concluded that  RSRV is not
actually dropping read requests (phew), but may be getting overloaded
and take an extremely long time to service the request.

This commit removes the retry logic, but leaves the timeout
configuration logic (and centralization) such that it is possible to
set a sufficiently long time (ex 10s) to avoid timeouts that will
crash the scan.