Support a Signal write_timeout pass to put.
This makes the timeout for a *write* configurable on a per-put and
per-Signal basis. It uses a similar pattern to what we did for read and
connection timeout in #788.

At SST, we were hitting a hard-coded timeout of 30 seconds (somewhere
in caproto or ophyd, I now don't recall...) and the motion of their mono
can sometimes take longer. With these changes, we could set
``write_timeout=180.`` in anticipation of lengthy motions.