Ensure that describe() works on ophyd.sim.hw objects.
It should be possilbe to call `describe()` any time after an object is
intialized. This fixes a regression from 1.3.x and adds a test.

We make an exception for area-detector-like objects until ophyd 1.6.0
applies a re-think to how `describe()` works on those to make it just
like everything else.

Thanks @lsammut for the bug report.