DOC trigger
Came to [this page](https://blueskyproject.io/ophyd/device-overview.html?highlight=trigger#trd) looking for a clear description of `trigger()` for a tutorial.  As it reads, the section does not connect the dots for a newcomer.  This change makes a quick summary and connects to existing documentation for more details.