API/FIX: don't raise AttributeError in _instantiate_component
If we do, the exception ends up being eaten by the
`__getattr__` (which have in place to make sure that multi-level '.'
access works in `getattr(...)`).

Instead, we now raise a RuntimeError along with a suggestion as to how
the user could have ended up in this state:

 - being a device sub-class (because we are in _instantiate_component)
 - having a Component at the class level that Python found and used as
   a descriptor
 - but the attribute is not registered with the Device/Component
   machinery.

A non-Device subclass as a mixin will definitely get you to this
state, but there may be other ways (monkey patching the class etc).

Mixing in a non-Device subclass used to work (by chance not by design)
in ophyd <= 1.1.