ENH: add a "busy" device
We had a request for a way to set a sleep in the RE that would give a
progress bar.

You can get an interuptable sleep via `bps.sleep`, but that provides no
feedback.  Rather than trying to invent another way to do progress bars, re-use
the progress bar machinery for hardware motion.