clarified source of an error message
in the put method to the EpicsSignal class, there is a check
to be sure that the signal can be written to, i.e. that
self.write_access is true.  that error message results in an unhelpful
traceback in that it provides no information identifying the PV
triggering the message.  in the case of a deeply nested signal, this
error message is unilluminating.  this commit adds self.name to the
error message.