Add initial orchestrators idea
Im not sure if the general ideia fits in ophyd scope, but this is an
initial attempt.
The idea is to have abstractions to perform tasks that are possibly
frequent to the plugins and drivers, but not basic enough to be put
in the initial plugin/driver abstractions. For example, a common
method to configure a grid of peaks in SimDetector or to one to
define how to average images in processPlugin.

I wanted to set a default behavior in the method but in a way that
parameter default values could be overriden by an user who knows what
they are doing.

Unfortunately, the only way i could come up to do this was by manually
coding the properties as possible keys to kwargs (eg. line 22 in
ProcessPluginOrchestrator). I tried dynamically checking if any argument in
kwargs was equal to one of the component_names but for that i needed to
use something like getattr(self.cam, component_name).put() and for some
reason that didnt work.