ENH: add ability to control component names are mangled
The Python name (via attributes) of a child device is `foo.bar.baz`, however
for a number of historical reasons (allowed keys in mongo and for attribute
access in Pandas data frames with columns named from the data keys in an event
stream) ophyd-sync has replaced '.' with '_' in the names.  This leads to
ambiguity when looking at the (default) name to which component on which device
the data came from as there may be `_` in the attribute names.  Ophyd-async
used '-' which does allow for round tripping.

This makes it controllable on per-instance basis in ophyd-sync so that users
can either make a different set of trade offs.

xref https://github.com/bluesky/ophyd-async/issues/666